Add relative loadpath scoped to the current file. Add module variable.
The module variable is a first step torwards node compliance. The
node.js module loading system assume that `module` is a reference to the
object representing the current module - e.g. `this`, and that
`module.exports` is the same as `exports` object. See
http://nodejs.org/api/modules.html#modules_the_module_object for
details. The test for this is currently commented out because it's not
functional yet.

Add the current module's directory path to the evaluation context's load
paths so that a module called superawesome found in
/usr/local/lib/node_modules/superawesome/index.js can do things like
`require('./lib/super.js')` and the file
/usr/local/lib/node_modules/superawesome/lib/super.js will be loaded.