Many additional test for Number.toFixed() and relevant code changes.
This also required some changes to NumberLiteralExpression.java to
handle numbers expressed in the scientific notation that is acceptable
by Javascript standards. For example, in Java you have to write 13.0E+21
while in Javascript you can write it this way, or as 13e+21.

DynNumber is a bit of a pain here. We have to check instanceof for that
vs PrimitiveDynObject and call getPrimitiveValue accordingly. This needs
to stop, @qmx, my friend. :)