Modify JavaFunction to handle method calls with varargs signatures.
Instead of always splatting the function call's arguments to individual
atoms, now we check the last parameter type to see if it's an Object[]
and if so, we don't touch it.

In general, the buildArguments call can probably be simplified or even
eliminated if we make better use of what's available through the
MethodHandle interface. For now, it's a clunky as you can see.