Disable the longer CoffeeScript test until we fix a compiler bug.
There's a place in the CoffeeScript code where it has a function like below:

        tag = (function() {
          switch (id) {
            case '!':
              return 'UNARY';
            case '==':
            case '!=':
              return 'COMPARE';
            case '&&':
            case '||':
              return 'LOGIC';
            case 'true':
            case 'false':
              return 'BOOL';
            case 'break':
            case 'continue':
              return 'STATEMENT';
            default:
              return tag;
          }
        })();

The bug I'm seeing is that the value of "tag" from inside the switch
statement's default case is retaining the value from previous
executions of this function. So, the first time this function gets
called and tag has the value of "CLASS" outside the function. On next
executions of this function, tag still has the value of "CLASS" inside
the function even though the value of tag is now something different
outside the function.