Move to dynjs.compile.mode={OFF, FORCE, JIT}
When compilation is JIT, code begins life as interpreted, and once a function
is executed 5 times, a request for async compilation occurs, and swaps out once
completed.

When compilation is FORCE, code is immediately compiled to bytecode.

When compilation is OFF, code is executed in interpreted mode.

At the current time, unless code is FORCE, invoke-dynamic integration for Java
code will not occur.  This needs to be fixed.

The current JIT threshold is 5 executions for a block.  Since basic-blocks are nested,
a function may compile its top-level portion, while the innards of if/else, try/catch
may not compile, as they each have their own thresholds which must be met before
compiling sub-blocks.