Tweak the bin/dynjs executable again.
This is another attempt to find a common ground between linux and the
mac OS. Mac OS doesn't like the -e flag to echo. The -e flag simply
instructs echo to interpolate escaped characters such as \n when the
string is enclosed in single quotes. To get rid of the -e flag, we can
use double quotes where escaped characters are automatically
interpolated.