Add unit and jasmine tests for implementing methods with booleans.
These tests are commented out because they fail. It seems that
jitescript has no problem generating the method signatures as booleans,
but when `new` is called on one of these objects, the boolean becomes an
integer. Jitescript method signature generation happens in the method
ObjectMethodGenerator#defineMethod. I believe that the problem is
probably somewhere in
JSJavaImplementationManager#createImplementationWrapper (but I'm not
sure).