Ensure we handle __proto__ when set in an object literal
When an object literal is created, we need to check for the __proto__
property and if set ensure that goes through our normal __proto__
handling logic. This feels like a hack to check for __proto__
explicitly but does get the included spec and the nodyn-examples
express example app working.

Suggestions for improvement are most welcome.